fix: Get_Last_Insert_Id now returns last insert id

+++ Server/AspNetCore/MonthlyLifeguardRegister/Classess/Database/SqLiteConnection.cs
@@ -122,7 +122,7 @@ namespace MonthlyLifeguardRegister.Classess.Database
 
             try
             {
-                return (int)sqliteCommand.ExecuteScalar();
+                return (int)(long)sqliteCommand.ExecuteScalar();
             }
             catch(Exception e)
             {
